Prepare for command costs, update NMS.

+++ commands/Fun/reddit.js
@@ -1,14 +1,15 @@
-const Command = require('../../base/Command.js');
+const Social = require('../../base/Social.js');
 const snek = require('snekfetch');
 
-class Reddit extends Command {
+class Reddit extends Social {
   constructor(client) {
     super(client, {
       name: 'reddit',
       description: 'Posts a random subreddit entry.',
       usage: 'reddit [-new|-random|-hot|-top] [subreddit]',
       category: 'Fun',
-      botPerms: ['SEND_MESSAGES']
+      botPerms: ['SEND_MESSAGES'],
+      cost: 1
     });
   }
 
@@ -23,7 +24,10 @@ class Reddit extends Command {
 
       if (!message.channel.nsfw && meme.over_18) {
         throw 'Cannot display NSFW content in a SFW channel.';
-      }  
+      }
+      const cost = this.cmdDis(this.help.cost, level);
+      const payMe = await this.cmdPay(message, message.author.id, cost, this.conf.botPerms);
+      if (!payMe) return;
       const msg = await message.channel.send(`'Fetching from ${meme.subreddit_name_prefixed}...'`);
       await message.channel.send(`${meme.title} submitted by ${meme.author} in ${meme.subreddit_name_prefixed}\nUpvote Ratio ${meme.upvote_ratio}\n${meme.url}`);
       msg.delete();
